requirements.txt
#PyCharm may remove discord.py~=1.7.2
discord.py~=2.3.2
aiohttp~=3.8.5
prettytable~=3.8.0
pyotp~=2.9.0
aiofile~=3.8.7
discord-pretty-help~=2.0.7
motor~=3.2.0
pymongo~=4.4.1
